+++ doc/BUILD.bazel
@@ -4,7 +4,6 @@ load("//bazel:python.bzl", "doctest", "doctest_each", "py_test_run_all_notebooks
 
 exports_files([
     "test_myst_doc.py",
-    "source/serve/doc_code/capacity_queue_request_router.py",
 ])
 
 # --------------------------------------------------------------------
@@ -305,7 +304,6 @@ py_test_run_all_subdirectory(
         "source/serve/doc_code/cross_node_parallelism_example.py",
         "source/serve/doc_code/llm/llm_yaml_config_example.py",
         "source/serve/doc_code/llm/qwen_example.py",
-        "source/serve/doc_code/capacity_queue_request_router.py",
         "source/serve/doc_code/capacity_queue_request_router_app.py",
     ],
     extra_srcs = [],
