+++ python/ray/serve/tests/unit/test_capacity_queue.py
@@ -4,16 +4,11 @@ from typing import List
 
 import pytest
 
-# isort: off
-# doc_fixture must be imported before capacity_queue_request_router
-import ray.serve.tests.fixtures.doc_fixture  # noqa: F401
-from capacity_queue_request_router import (
+from ray.serve._private.experimental.capacity_queue_request_router import (
     CapacityQueue,
     CapacityQueueStats,
     ReplicaCapacityInfo,
 )
-
-# isort: on
 
 
 def _get_raw_class():
